+++ Tests/StoreTests.swift
@@ -28,7 +28,7 @@ fileprivate class TestObservable: ObservableObject, AnyObservableObject {
     var notPublished = 0
 }
 
-@MainActor class StoreTests: XCTestCase {
+@MainActor final class StoreTests: XCTestCase {
     override func setUp() {
         super.setUp()